fix: RecursorVal.getInduct to return name of major argument’s type
Previously `RecursorVal.getInduct` would return the prefix of the
recursor’s name, which is unlikely the right value for the “derived”
recursors in nested recursion. The code using `RecursorVal.getInduct`
seems to expect the name of the inductive type of major argument here.

If we return that name, this fixes #5661.

This bug becomes more visible now that we have structural mutual
recursion.

(The implementation is still up for discussion; let’s first agree on
whether this is the right semantics.)

(Also likely worth renaming this function to `getMajorInduct` for
clarity.)