fix: verso docstring {name} role suggestion overload (#10663)
This PR disables `{name}` suggestions for `.anonymous` and adds syntax
suggestions.

When the provided name can't be resolved, the `{name}` role suggests
fully-qualified variants. But if the name is a syntax error, it
attempted to suggest names that had `.anonymous` as a suffix; the
resulting list of suggestions of all names in Lean's environment
overloaded the language server.